requirements.txt
pyyaml
django-river
